Let me resume.
Actually, by any means this is not easy, not even for me with all my experience.

I've decided what should be done before the input is passed in and how to convert the stuff to one hot vectors - I'll just use the 0-127 range for the inputs and the outputs, but in the outputs I am a bit undecided. I also do not understand the thing with the temperature.

There is really a frightening amount of detail to go through before I can get my first net running.

Before I start I need to decide what should be done with the output.

Yeah, there is the softmax - then after that is some kind of selection function necessary?

And while doing inputing the previous output should I apply that selection function and feed it in, or should I do the softmax as is?

I will have to implement the softmax as well.

Ohhhh, what a pain in the ass. Trully what a pain in the ass.

Why is it my responsiblity to drive ML library development on Windows? Why is Python + Linux by far the dominant platform? Why are all other functional languages so dogshit for systems development?

Well, look at the bright side - I have Spiral now.

In six months this will all be behind me and I will have a stable platform to play in. I have only myself to blame for sticking to my principles.

Make no mistake - finishing the last 20% of the ML library and then retracing my steps from last year until I get over the hump will take some serious time.

That having said, I can't start programming until I have it worked out completely. Let me dwell on it for a while longer. I need to study more.