I haven't done any coding yet, but I think I have a plan formed on how to deal with the outputs.
In 2015, I did a lot of experiments with sparsity based activation functions so my memories of that are clouding my judgement. The way Karphathy does it is that he does not propagate gradients through the input.

I probably don't need to do that kind of thing anyway. I am dedicating way too much mental effort towards making that useless bridge and it probably even if I implemented that, it would not improve training. No doubt ML researchers investigated turning the sampling function into an activation already - but I will try it at some point.

With that I can narrow my aim a little.

https://youtu.be/iX5V1WpxxkY?t=1217

On the CPU, doing the sampling is easy enough. I am looking over the Lua code and am not exactly figring out how that is done there.

But it does not matter. To do that kind of thing on the GPU, I'd need a scan + a reduce.

Agh, how troublesome.

I haven't done everything I need just yet. I am going to have to take a detour and spend some time working on that.

Also in preparation for the optimizing the LSTM cell, I am going to have to figure out how to fuse a scan + reduce where applicable.

Actually, quite many operations are fusable when done on a block by block basis.

It is really a pity that Cuda does not offer any device synchronization abilities inside the kernel.

It is also a pity that it resorts to using those shitty streams for concurrency when batching operations is a better choice 100% of the time.

Getting back on track - I need to do the damn scans.

This is something that I studied over and over and various contexts and never quite figured out how to do effectively. What I mean is that unlike the reduce operations, I failed to internalize it.

Also it might be time to maybe start breaking my dependency on Cub. Since I cannot pass in variables inside closures, that really leaves me with much less reason to use those things.

Here is the current status of my char RNN so far:

1) The prepass is planned out. It won't be a problem to implement. There are various ways of chunking data into tensors and preprocessing it on the GPU. I will make the right choice here.

2) As strange as it sounds, the actual RNN won't be a problem either. There are some problematic steps around having to add regions in order to curate the output and also preserving the state while doing the truncation - the lack of GC is really screwing with me here.

Maybe I will bring back the lazy values in reduce in order to simplify things, but I want to try the new ideas that I have.

The most difficult part will definitely be figuring out how to preserve the state - but I will be able to manage it.

3) The damn sampling. This will not be an insignificant amount of work. Both doing the scans and the selection functions will push me to my limit.

But doing this was inevitable. On the CPU scan is not particularly important, but on the GPU, it is greatly important.

I cannot really say that I understand GPU programming until I've internalized the various ways of doing it.

---

So #3 which is the most diffucult part will have to be done up front. I will dedicate the next few days to implementing and internalizing scan.

Now, before I do that let me just look a little bit on how sampling is implement in PyTorch and then I will move studying the algorithm. I think I should probably look into how Cub does it as well.