I am up. As per plan, let me study the Cub library for a while and then I will do some programming.
I've been thinking about optimizing the LSTM and as expected, the new tensor class will make it much, much easier. It is really great that I figured out what those leading dimension parameters in gemm are as well. That will be really convenient.

I really needed Spiral for this. Types in F# are just too rigid for this kind of thing.

Now...let me slowly start. I need to study Cub and then I will over time slowly integrate its functions directly into the Spiral library.

Today, my task will be to do the basic d2 stream and use that to implement the sampling function.

Regardless, I won't be extending the language anymore, especially not with functors in order to do iterators. That is way too complicated. I'd rather put in the extra effort that will pay in the long term by breaking the dependency on Cub.

That takes a load of my mind.