https://stackoverflow.com/questions/6404992/cuda-block-synchronization https://devblogs.nvidia.com/cuda-9-features-revealed/ http://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#cooperative-groups
Just last night I was thinking that Cuda really needs some way of synchronizing blocks. Yeah, this decides it for me and is another load off my mind.

Rather than trying to figure out how the hell threadfence works in order to synchronize the blocks, I'd be better off buying a new GPU sometime in the future and changing the scan algorithm to take advantage of cooperative groups.

For now I think I will opt for a much simpler 3n scan algorithm rather than use the hack from Merril's paper.

Trying to work out how to make the memory ordering exact with a bunch of hacks is not worth learning.