Wow, cooperative groups are absolutely game changing for Cuda programming. Yeah, I won't bother optimizing the scan since I will be tearing it down later anyway.
ManagedCuda is still at 8.0 anyway so that is the only silver lining that I have currently; I would be able to take advantage of it because of that. I could use it inside kernels, but I do not think ManagedCuda has the ability to launch kernels with cooperative groups right now.

...Ok, now where was I before I started studying cooperative groups?

Yeah, I still want to go through Cub from top to bottom. After I am done with that, I will be ready to start on making the scan kernels.

Slowly, the pieces are falling into place for me.