I am just about ready to start let me gather my thoughts first. I'll lay them out and build a plan based on the stuff I've studied.
First let me start of with Cub.

After studying its source - though it is well engineered for a C++ library, it is still a C++ library. Even if Merril was the absolute best in the world, there is only so much he can do to stop the thing from looking like it crawled out of sever.

Furthermore, there are some composability issues with it.

For example, I'd want to fit in a map at the end of the device scan. That can't be done, not without launching a separate map kernel.

For the block scan, I'd want to use it to implement a chained scan.

How exactly should I broadcast the returned value? Well, I can make a singleton shared array, have the last thread write into it, then do block sync and have all the threads read from it.

Now instead of doing that, why not just load from the shared array?

Now I could do this with Cub, but what guarantee do I have regarding how it will use the throwaway memory?

Guarantees require personal implementations.

I do not require backwards compatiblity in my library, and Spiral is capable enough of supporting the necessary primives on its own.

So I might as well make them for both scan and reduce and get those things out of the way. I'll want to benchmark my own implementations against Cubs in order to ensure that I am not leaving too much on the table.

The second thing worthy of mention are the Cuda cooperative groups. I only found about them like yesterday.

Their existence is absolutely game changing.

Not only would they trivialize the implementation of the fast 2n chained scan and greatly simplify other implementations - their impact goes greatly beyond that.

Forget kernel fusion - I could literally fuse entire forward and backwards passes into just two kernels!

Now that would get the function call overhead out of the way!

Well almost. I could really use the block level gemm and convolutional functions in order to get all the way there, but if it just matmult, I could probably figure something on my own.

So here is my plan - I'll make the optimized LSTM to the best that I can with my current abilities, post in on the ML sub and try to bait Nvidia into sending me latest Titan so I could do the fully fused network.

Well, that is nice dream, but me and social success do not exactly mix, so I won't really bet on such a thing. Regardless, it would be a good target to aim for since I was going to do it anyway. Free upside is always nice.

Now...the only thing left would be to make that scan. I'll use a shared memory Kagle Stone algorithm for block reduction and aim to do the outer level in registers first...well, actually let me leave that out for the first version. I should aim for simplicity on the first try.

Now I do not expect I will be done with this today given that it is already 4:35pm - I'll stop for the day in an hour, but let me do some mop up.

In `map_d2_redo_map'` that padding thing has been bothering me. Back then I had forgotten how shared memory bank conflicts worked, but now I have a firm understanding of it after a refresher. I know how to deal with it.

Let me clean that up and then I will work on the scan. It won't be hard.

The very real work that I did was all the thinking and studying during the last few days - what I will from here on is just mop up.

Also unlike before, I won't have do do with horribly vicious debugging. In all likelyhood that is behind me so I will be able to make smooth progress.

It is time to start.