Here it is, the completed d2 reduce kernel.
It really wasn't a big deal once I worked out the kinks. The knowledge I've gain will form the basis for doing the scan kernels.

There is no mistaking it, with this I can break the depency on Cub's reduction and scan functions. I'll keep it around for a while longer though.

Let me stop here for a bit and I'll do the scan after that.