Let me finally start for the day. I got up really late today and did not really feel like starting.
https://www.youtube.com/watch?v=7HKbjYqqPPQ

And when it was time to start, I'd figured I'd rather have this.

Right now I seem to be hung up on how to remove the neutral element from the reduce operations - forget it.

Let me try finishing the scans today. First let me add another in' to the mix.

Then I will make the auxiliary function that takes advantage of the allocator for the output and then.

Then I will do a scan that does it over the outer dimension.

It is time to start.