I did need to debug the d1 scan a little as it turns out, but it works fine now.
Ok, great.

Now let me make the d2 scan as well though I do not really need it.

I just want to deal with this today. After that I will make the device wide scan. That one will be a pain in the ass, but I need to do it.