The truth is, I am wondering whether that if is necessary or whether cub will know to coalesce the threads properly on its own. I think cooperative groups are the answer to this. With CG I could probably get rid of the neutral element for good.
Actually, I should try the reduce with half the threads off to see if Cub will do the right thing. Two days ago while I was rifling through its source, I did note that a lot of its complexity is keeping track of active threads.

I've been thinking of finding a way to remove the neutral element, but this is the first thing that should have occurred to me.

I do not feel like it anymore. My energy is spent for the day.

The device scan is finished. I'll test it tomorrow along with the idea I've mentioned.

Today I did quite a bit.