Yeah, nice. The scan is now operational. I will admit that the way Cub treats aggregates in the inclusive scan is confusing, but that is why testing is for.
I guess I will take a break here.

Now that I have this, the time is to figure out how to implement the sample function.

I'll probably have to extend the scan so the initial map is passed in an index.

...

Ok, I have it. I need `mapi_d1_inscan_reduce_map` for the sample function in order to compact it all into a single kernel.

Yeah, that should be it.

I'll implement it after the break. Slowly, the Cuda library functions are coming together.

You'd expect a great ML library to have a rich library of GPU primitives and that will be the case in Spiral.

Since I've spent the last 30m in though, let me stop here.