The thing works now.
Great, excellent.

`kernel10` is pretty much the sampling function.

I am going to need to make another kernel for the softmax. It is map + redo + map, but in a different direction.

Well, I will keep that for the last. It won't be hard to do. I am definitely getting a lot of exercise at GPU coding here.

If I am right, I think that layer norm is similar to the softmax so that should come in handy there as well.

Let me take a break here.

I need to think of what should come next anyway.