"2pm. I am still thinking about the transpose. Yeah, there is no way around it, accessing the transposed values would be convoluted, but at least I fully understand the padding rules.
Also I was right, the reduce code by Harris does have data races in it.

2:10pm. Ok, let me skip to lesson 3. I know all of this already.

2:15pm. Focus, stop surfing /a/.

2:25pm. Ahhhh, goddamn shared memory banks. Yeah, I thought I figured it out, but then if the number of items per threads turns out to be 32 or a multiple of it, I'd end up having the same damn problem.

So not only do I need to pad the right side, I somehow need to pad the items per thread as well so it is not a multiple of 32.

4:15pm. I had to go lie down to think and now I just finished the chores.

The whole day it seems, I've been thinking about how to do the transpose. I understand now how to do it without suffering bank conflicts, but nevertheless, the block transpose method I have in mind would be limited by the amount of shared memory available. It is not a full solution.

I won't do any programming today it seems.

At any rate, I did figure out that in the kernel I want to make, I do not actually need the damn transpose. All I need to do is load it into local memory. So that solves one thing. I am going to have to take advantage of the Cub functions for this.

I want to continue watching the lectures.

4:35pm. Ah, I understand the exclusive scan. I know how I would implement it. Really, all I need to do is to make the initial element the first entry in the shared memory array.

Well, that is one mystery down.

I mean, it seems obvious, but there are other ways of inserting the first element and I was not sure which one should I pick. This one definitely.

Now...

Let me see if I can internalize the other scan algorithm. Blelloch's.

4:40pm. https://classroom.udacity.com/courses/cs344/lessons/86719951/concepts/880351240923

Wow, I get it completely now. The key is the mirror image. The algorithm could be adapted to more efficient stride patterns.

4:55pm. I am looking at the other algorithm by that Chinese guy, and unlike the Blelloch one, I don't really see how to make it parallel. No really why does he explain the inefficient pattern?

```
__global__ void work_efficient_scan_kernel (float *X, float *Y, int InputSize) {
    __shared__ float XY[SECTION_SIZE];
    int i = blockIdx.x*blockDim.x + threadIdx.x;
    if (i < InputSize) {
        XY[threadIdx.x] = X[i];
        }
    for (unsigned int stride = 1; stride < blockDim.x; stride *= 2) {
        __ syncthreads();
        int index = (threadIdx.x+1) * 2* stride - 1;
        if (index < blockDim.x) {
            XY[index] += XY[index - stride];
            }
    }
    for (int stride = SECTION_SIZE/4; stride > 0; stride /= 2) {
        __ syncthreads();
        int index = (threadIdx.x+1)*stride*2 - 1;
        if(index + stride < BLOCK_SIZE) {
            XY[index + stride] += XY[index];
            }
    }
    __ syncthreads();
    Y[i] = XY[threadIdx.x];
}
```

As expected the thing on page 209 has data races because it does not use temporaries. So much for the thing in the book. It also neglects to mention raking by the outer dimension.

This is something that is easy to do in Spiral as it allows passing state from iteration to iteration, but is difficult to do in C++.

5:05pm. The way indexing is done here is quite convoluted and it is difficult for me to understand what it exactly going on here. There is no padding of shared memory arrays either, and the access pattern will lead to bank conflicts.

In 2016 I could not really see why the Udacity course skipped over this algorithm completely, but now I understand why. The guy who wrote this is an amateur. To think that this was my first introduction to parallel programming."