I am done for the day. Damn all this stuff as simple as it seems was actually quite hard. I sure am tired.
Today I did the change to whilecd so it uses the HostTensor, I refactored the cub functions and I even finished the softmax forward activation.

Well done, well done...

Tomorrow I will look into what the backwards pass for the softmax is supposed to be and after that I can seriously start to consider going forward with char RNN.

Char RNN, layer norm and then I am seriously going to dive into reinforcement leanring. No doubt a lot more GPU programming will await me along the way.

But actually that is good.

In terms of richness, the current Spiral library is already starting to vastly exceed the previous one.

By the end of the next month, it will become of obvious just how far I have come.

The power that I will attain is waiting for me in the future and this effort I am putting in is just a step closer to it.