I renamed the replicate function to d2 and am almost ready to start the day.
I am still thinking what should be done.

Well, the softmax. Should I try to do it all at once or should make the function piece by piece first?

I'd need to make d1_replicate for that.

Well, I should make that function anyway. I also need to refresh my memory how the adjoint for the backwards step of division is calculated.

...

https://eli.thegreenplace.net/2016/the-softmax-function-and-its-derivative/

Ah, this is is bringing me back. I remember doing the derivative for the softmax in the Hinton's course and this is a great refresher.

I never really figured out the equations all the way, but that won't be necessary.