Let me slowly start the day.
These kernels are kind of a grindfest that won't end, but eventually I will get to actually playing with the char-RNNs.

1) Could I by ditching Cub improve performance due to taking advantage of Spiral's toa tensors?

2) Since many of the kernels have statically known sizes, should I rewrite them so to make it easier for the compiler to unroll them or is the compiler already decent at this?

The only thing that occured to me last night are these two tensors. But I do not want to ask them online right now.

...

Ah, I see. I actually forgot that for the indexing kind of cost, I can already do it to an extent through the d1 redo kernel. Instead of converting the labels to one hot vectors of float32, I can just keep them as unsigned ints.

That damn Python code from yesterday really inpired me. For the softmax cross entropy, I'd really want to be indexing using the labels into the matrix.

...Ah, yes, I think I have a bug in reshape in that I am not really taking account that the dimensions might be padded when calculating the length.

I need to take this into consideration.

I will fix it immediatelly. Also I need to redesign them a little. Calling HostTensor function is such pain in the ass, ideally I'd want as much as possible to be in the facade.

Well, I need to think a little about this one first though.

One thing that I am sure of is that for the time being I do not feel like working on the kernels anymore.

Starting from here, I think I will assume that I have everything in place and move forward until I actually need the missing pieces.

So today I will make it my goal to load the dataset and pass it through the system. Then I will begin work on the RNN combinators. It won't be too hard.

For now, let me fix the reshape.