Let me stop here since lunch iwll be in a bit.
"I had an idea.

If multiplying `[a,b,c] x [c,d]` then I could just say it comes out as `[a,b,d]`. That way on the first input I could fold all the steps into a single large multiply rather than 64 of them.

I had an idea. Reshape as it is right now sucks donkey balls.

For a tensor `[a,b,c]` why couldn't I simply pass in an inlineable `inl a,b,c -> a*b,c` in order to merge the dimensions. I am really going through so many complications right now. If the `from` field of the module is 0 then I could just pass in the `near_to` reflexively."

I am going to work on reshape. Maybe view should get the same treatment.

At any rate, this manual fetching of dimensions needs to end.

Also for a while now I've felt that having lower and upper boounds for dimension is worthless. It is just creating complications for me and is a bait for inefficiency.

Right now I am tempted to modify the arithmetic operators to support operations over {from near_to}.

It is just, the cases where I actually want the lower bound to be anything other than 0 have been exceedingly rare so far.

That settles it. I will simplify the tensor radically by removing the unused feature that imposes itself everywhere.

That will allow me to make view and reshape a lot more elegant.

I am done for the day.