I got up really late today.
http://blog.benoitblanchon.fr/safehandle/

Definitely, today will be dedicated to me learning about SafeHandles. I am going to be following up on the SO answer from yesterday by msedi.

I am going to have to think a little of how I want to deal with this in Spiral given that I do not support inheritance in order to implement safe handles.

But let's keep that aside for until I am done with studying them.

After I am done with this, I will do what I had planned yesterday which is refactor the HostTensor so as to remove the bounds from it.