I am done with studying `SafeHandle`. I understand what it does, but it is no wonder I was not comfortable with them over a year ago when I first found out about them. The information on them is rather sparse and my understanding of them is shallow. But they should be doable. Being said that they work by someone else is encouraging. I won't be able to use them for allocating arrays, but maybe for managing contexts they will come in hand.
That having said, I was not counting on them being so non-generic. I won't be able to use them directly in Spiral, so I think I will put them on the backburner for the time being.

This is because they need to be inherited from in order to be implemented. Maybe I could do a generic implementation in F# and pass in the release as a lambda? Bah. Forget that for a bit. It is not important right now.

I'll go with my current memory management scheme for now.

Let me take a break and then I will start redesigning the tensor.