Ok, I am done with the refactoring. I am too tired to continue past this point.
Is this a simplification or not? I am not quite sure anymore.

I guess I'll give it a few days like this.

Well, forget about that. I'll get the benefit of simplifying things in the long run.

What was I even supposed to do past this? The init kernel?

Ahhh...let me just close for the day. My stress is at its limit. I'll do much better once I am fresh.

Today's refactor was rather tedious. This is a contrast to yesterday when I was in the zone the whole time.

Hopefully I will get back into the swing of things tomorrow.

I just need to take it easy.

I'll figure out how I want to load the models and then get to work on making the RNN segment. This will take me a week maybe.

Just think about it - I'll finish the char RNN in 1-2 weeks and then I will have everything I need to go into reinforcement learning.