Pfffuuu. I am done. Definitely not going to run this unless I want to kill myself today. It is probably all broken unless by some miracle I got it all right the first time around.
I'll test this tomorrow and then move to making the `init`.

All in all `grid_for` is a great idea that I should have done from the start. If it wasn't for NVCC being shit, I'd have managed to root out the `neutral_elem` from kernels that do not need it.

The things I am doing today is why Spiral was made. What I am doing here is functional programming at its finest.

I was completely in the zone today and now it is time for my well deserved rest.