"9:45am.
```
Error trace on line: 6, column: 16 in file "Core".
inl error_type x = !ErrorType(x)
               ^
lit The inner sizes of the tensor being reshaped must be contiguous.
```

Only this thing is left. Yeah, this is wrong. I probably fucked up the check somehow. Let me take a look.

```
assert (Tuple.tail size = Tuple.tail size') "The inner sizes of the tensor being reshaped must be contiguous."
```

How about I compare the tail?

9:50am.
```
            inl make_body dim = make_body
            inl _ =
                inl {size} = make_body dim'
                assert (Tuple.tail size = Tuple.tail size') "The inner sizes of the tensor being reshaped must be contiguous."
```

I am taking out this check. The scenario I am trying to guard against can't be guarded like this.

```
assert (Tuple.forall ((=) 0) o') "The inner dimensions much have offsets of 0. They must not be 'view'ed. Consider reshaping a copy of the tensor instead"
```

Also I am checking for the wrong thing here.

10am. Ergh...

10:30am. I've been thinking about this for the last half hour.

What I am going to do is simple. Since it is impossible to support `reshape` in a way that would be correct, I am going to remove it from the language and replace it with `split`.

I am also going to remove `offset` and replace it with a scalar of the same name. That will allow me to fully support rotated and non-contiguous tensors.

Let me do this and then I will work more on the loops in order to root out the neutral element.

This is actually a fairly major design change for the tensor.

It should have occurred to me before that reshape does not make any sense, but I've been affected by it being there in Numpy and Tensorflow so I tried to make it work."