Today reshape gave me a lot of trouble, but now that I did flatten the way to do reshape comes out naturally from combining flatten and split.
Trully this is what I look for in programming. Thanks to this I won't have to redo the map kernel...

That having said, I really in fact should redo it in order to support kernels that have been viewed in the inner dimension.

But now forget that, let me just get this to work if possible and I will take a break.