Well, I think this should do it for init.
What I had yesterday was equivalent to setting the thread limit to 1 for a given dimension. It really did not feel right to let a single thread iterate over a 128 sized line. A limit of 32 should be ideal.

Meh, let me take a break here and I will move to testing after that.