Init works and it works very well. This was a piece of cake. It gave me no trouble at all.
I guess now it is time to put it into operation.

It just occured to me that I cannot put in num_steps as a dimensions because otherwise I won't be able to get the label for the last step.

Well, let me think about the subject now that I've managed to come thus far.