I hate it when I am pissed off and the computer picks that time to play tricks on me. Probably the out of memory error had nothing to do with what I did yesterday. All memory allocations go through the allocator anyway.
But this is not right.

I can't fly off the handle as soon as I run into a limitation.

But all the ManagedCuda types which keep getting in the way, the error checking which I can't emulate in Spiral and which would require extensions to the language just for this single purpose, and the opaque classes whose functionality is duplicated hundreds of times do not feel right to me.

I am going to purge CUdeviceptr and the rest of that nonsense from the library.

I'll only it and SizeT and transfer boundaries when I need to.

Seriously, this is not that difficult. Wasting days and getting angry at a simple array copy is not something that should be happening to me.

But my instinct when programming is to make everything right and this stupid thing has just resisted and resisted me to no end.

https://devblogs.nvidia.com/unified-memory-in-cuda-6/

No more. I've noticed just now that there is a simple memCpy function for all cases. I will rewrite the way pointers are handled, how tranfers are done, and I will look into managed memory.