I've undone the changes.
Bah.

This is a good lesson for me. Yeah, sure I've proven the the neutral elem could be decisively removed in various kernels. It just did not occur to me that it would not be a good idea even if I could. Simplifying it trivially for the user is not worth it for the library.

So I am done with what I wanted to do yesterday. I got the transfer functions and the grad checker in order and even fixed that compiler bug.

So what is next?

Well, obviously loading the dataset. Just how long is it going take me to do that thing? After that I can finally start work on the RNN segment of the library.

I need to some time to gather my motivation for the next part. I've gotten led astray way too much by the various things. Let me take a break here for a while.