The thing that definitely bothers me most of all would be the variable filtering thing in the language.
I did not do it right. The two bugs are proofs enough.

Instead of delaying the filtering, I should make it so only useful work is done.

That means instead of iterating through the whole used variable set, I should reduce them so as to take the diff from step to step.

I am imagining them as filtering nodes. That I put the filters into the functions directly confuses things somewhat.

And I need to specially handle `open`. `open` should always be filtered directly.

Yeah, this is easy. I do not even have to change particularly much for this to take effect. I just need to add an extra thing into the prepass.

Since open is so important, as important as the functions, I think I am going to add it as a keyword to the language so as to allow it to be filtered. But really, I should do the damn nodes much like position.

...Once I finish touching up the language, at that point I think I can finish loading the dataset and moving to making the RNNs. It really has been too long.