Let me run this now.
```
The timings are: (00:00:05.9880045, 00:00:05.8884744, 00:00:03.2834406, 00:00:00.2724015)
The time it took to run all the tests is: 00:00:15.4323210
```
This is for Learning. Wow surprisingly it is dominated by parsing and the prepass. Peval is actually quite faster than them. Let me check out the standard library tests.

```
The timings are: (00:00:04.5354557, 00:00:04.6455406, 00:00:01.3895796, 00:00:00.1470141)
The time it took to run all the tests is: 00:00:10.7175900
```

Er, let me also add them up. Done.

Now that I look at the above, there is no point at all in even trying to speed up the peval. But I can't believe this, how come partial evaluation is like 2-3x faster than the damn prepass?

Let me also make these changes in master so I can get the timings for the previous version. Probably the prepass will turn out to be the culprit. If it got 2x slower that could easily explain why compilation got slower by 3s...well, not really, but...