``` The timings are: (00:00:04.6867997, 00:00:00.9593290, 00:00:01.3848872, 00:00:00.1498608) The time it took to run all the tests is: 00:00:07.1808767 ``` Uoh. Well 7.2s is uncharacteristically fast. Usually it is more like 7.6s. I really made the prepass quite a bit slower than I'd expected.
Let me try out `Learning`.

```
The timings are: (00:00:05.9014200, 00:00:01.6521766, 00:00:03.5617702, 00:00:00.2749159)
The time it took to run all the tests is: 00:00:11.3902827
```

I am testing why the changes in the other branch made compilation so much slower. Yeah, the prepass is the problem.

I need to consider how I am going to optimize it.