"8:45pm.
```
inb input =
    inl data = CudaTensor.to_dev_tensor data
    CudaKernel.init {rev_thread_limit=32; dim=seq,minibatch,size.hot} (inl seq minibatch ->
        inl x = data minibatch seq .get
        inl hot -> if x = to uint8 hot then 1f32 else 0f32
        )
```

Strange error my ass.

```
inl minibatch,seq = data.dim
inb input =
    inl data = CudaTensor.to_dev_tensor data
    CudaKernel.init {rev_thread_limit=32; dim=seq,minibatch,size.hot} (inl seq minibatch ->
        inl x = data minibatch seq .get
        inl hot -> if x = to uint8 hot then 1f32 else 0f32
        )

input
|> view (const 4)
|> CudaTensor.to_host_tensor
|> view (inl a,_,c -> a,4,c)
|> HostTensor.show
|> Console.writeline
```
Jesus I finally got the type errors out of the way.

8:45pm.
```
|> HostTensor.show
|> Console.writeline
```

I've written this particular segment so often that I might as well put it into the HostTensor module.

8:55pm. Ok, enough of this. I am done for the day. The dataset is loaded. I'll think of the optimal way to traverse it tomorrow. Since 500Mb is not too much for a dataset maybe I will keep it like this. For larger datasets I might want to prepare the second batch while the first is being processed.

There are all sort of parallel programming problems that I will need to tackle with Spiral. Really, memory management is such a mess and I do not have a good solution for it.

I am going to have to make some kind of region swapping system for iteration to iteration. I won't actually need it for the outputs surprisingly enough, it occurred to me that the strides between the outputs should be even with stack based region allocation so that is a relief, but I will need a region swapping system to deal passing state.

There is something I do not understand though. Why did the non-to_dev_tensor did not get caught by the blittable check?

```
{bodies={ar=var (layout_stack <function>); block_toa_map=[]; offset=lit 0i64; size=[lit 8714i64, lit 1i64]}; dim=[{from=lit 0i64; near_to=lit 128i64}, {from=lit 0i64; near_to=lit 8714i64}]}
```

Hmmm...

9:05am.

```
    print_static (blittable_is data)
```

It says this is false so what the fuck?

9:05pm. Ah, I get it. This is during the type check.

At any rate, I am going to call it a day here. Good job putting a little extra. I knew that loading the dataset was not a big deal. I'll figure out how to parallelize everything later. I'll do that while optimizing the LSTM with streams.

Forget that for now. For now I will just get to converge on a single short sequence."