Let me just start this thing. Spent my morning watching a talk instead of programming.
I guess I really am scared of starting this for some reason, but that work cannot be delayed any longer.

Focus. It is time. This is why I've been working on a language for over a year.

Forget about the cute optimization tricks and just do it straightforwadly. Forget the softmax and do cross entropy. Do the single step.

One by one I will scale it from the to an optimized char RNN with multiplicative integration and layer normalization.

I will prove Spiral's worth with my own two hands.