I am just about ready to finally start. I am still not sure of the right way to handle recurrent state. There are various ways and for starters I will go with the simplest functional way.
If I require more flexibility later on I will just rewrite the library.

It is not really a problem.

But I did need to put in a decent effort it trying to figure this out.

Now since I've spent half the working day resting in order to get my vitality back where should I start.

For the time being, what I will do is pretend this thing is a feedforward net and get it to converge on a single input. Then I will move to a single sequence.

Then I will move from there.