Something occurs to me, isn't map fold just a scan?
Now that the moment of truth is here I am starting to get confused by the complexities of all of this.

Instead of trying to squeeze this into one operation I will do a `map_scan` just to propagate inputs. Then I will do a monadic map to actually execute the operations. That way things will make sense.