This is a failure.
```
        inl wavefront network {input label} =
            map_fold (inl output x ->
                match x with
                | {rnn} -> {x with output label}, ()
                | {rnn output=output'} -> {x with output}, output

```

No, what I am trying to do here is code in by hand a very specific form of a generalized idea. The cost is essentially blowing the thing I had in mind into smithereens by making it too complex. I can't do it like the above.

Nothing I had in mind yesterday was wrong, but my approach to the problem is just too weak. In fact I do not need to do this at the level of layers!

I need to do this at the level of monadic steps, that way I could generalize this in its entirety.

Let me take a break here as I need to do the chores. I need to think about this anyway. I need to go at this from the ground up.

Since I am making this library for recurrent nets, I should take this opportunity to redesign it in order to parallelize it in its entirety.

https://docs.microsoft.com/en-us/dotnet/fsharp/language-reference/computation-expressions

It is not like this is a wild idea. The `eventually` monad is a variant on what I need here. If I start from scratch, I could probably get to the full thing.